Remove unneeded optional unique_ptr wrappers
Return just the unique_ptr<Expression> rather than wrapping that in an
optional.

We can signal the exact same thing as a nullopt by just returning a
nullptr for each of the functions that parse an expression, so removing
the unneeded wrapping makes the code a less verbose and more
readable.